+++ src/routes/VehicleRoutes.ts
@@ -3,19 +3,23 @@ import { IReq, IRes } from './common/types'
 import HttpStatusCodes from '@src/common/HttpStatusCodes'
 import check from './common/check'
 import Vehicle from '@src/models/Vehicle'
+import Authorization from '@src/util/Authorization'
 
-const getAll = async (_: IReq, res: IRes) => {
+const getAll = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAdmin(req, res)) return
   const vehicles = await VehicleService.getAll()
   res.status(HttpStatusCodes.OK).json({ vehicles })
 }
 
 const add = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAdmin(req, res)) return
   const vehicle = check.isValid(req.body, 'vehicle', Vehicle.isVehicle)
   await VehicleService.addOne(vehicle)
   res.status(HttpStatusCodes.CREATED).end()
 }
 
 const update = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAdmin(req, res)) return
   const id = Number(req.params.id)
   const vehicle = check.isValid(req.body, 'vehicle', Vehicle.isPartialVehicle)
   await VehicleService.updateOne(id, vehicle)
@@ -23,6 +27,7 @@ const update = async (req: IReq, res: IRes) => {
 }
 
 const delete_ = async (req: IReq, res: IRes) => {
+  if (!Authorization.isAdmin(req, res)) return
   const id = Number(req.params.id)
   await VehicleService.delete(id)
   res.status(HttpStatusCodes.OK).end()
